// types/model/bookmarkResponseDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * API 문서
 * Walkavel API 명세서입니다.
 * OpenAPI spec version: 1.0.0
 */
import type { LandmarkSummaryDto } from './landmarkSummaryDto';

export interface BookmarkResponseDto {
  /** 북마크 ID */
  id: number;
  /** 사용자 UUID */
  userId: string;
  /** 콘텐츠 ID */
  contentId: number;
  /**
   * 생성일
   * @nullable
   */
  createdAt: string | null;
  /** 랜드마크 정보 */
  landmark: LandmarkSummaryDto | null;
}